pyproject.toml
[project]
name = "alred"
version = "0.1.0a1"
description = "Collect LLDP and running-config, normalize links, generate containerlab, Mermaid, Terraform, and VNI outputs, and push config to devices"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
keywords = ["network", "lldp", "containerlab", "mermaid", "terraform", "cli"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: Apache Software License",
  "Environment :: Console",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Networking",
]
dependencies = [
  "Jinja2>=3.1.4",
  "netmiko>=4.4.0",
  "PyYAML>=6.0.2",
  "python-dotenv>=1.0.1",
]

[project.urls]
Homepage = "https://github.com/suzuyu/alred"
Repository = "https://github.com/suzuyu/alred"
Issues = "https://github.com/suzuyu/alred/issues"

[project.scripts]
alred = "alred.cli:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["alred"]

[tool.hatch.build]
ignore-vcs = true
include = [
  "/alred/**/*.py",
  "/alred/j2/*.j2",
  "/alred/sample_configs/*",
  "/README.md",
  "/CONFIG.md",
]

[tool.uv]
dev-dependencies = []
